Make do-upload params more similar to uploader's
Related ticket: [CODE-2611](https://codecovio.atlassian.net/browse/CODE-2611)

We want to make usage of the cli more similar to the uploader for the `do-upload` command. One way to do that is to accept arguments in a similar fashion the uplaod does.
These changes add more ways of passing arguments to the `do-upload` command.