Fix logging of tests
Some tests depend on the logging output. A problem we were having, most likely due to the fact that `CliRunner` is global AND that we use the same logger everywhere (`logger = logging.getLogger("codecovcli")`) was that multiple handlers were active for the loggers during tests. This caused the output to be repeated and assertions to fail.

My best guess is that `pytest_configure` runs for every test file that is executed, and the `CliRunner` re-uses the same logger for the isolated containers (which are not so isolated after all).

So the workaround is that before running tests that depend on output we double check that we have just 1 active handler.
The real real solution is probably to use different loggers for different files.