Change default uploader + refactor upload service.
The objective change in this commit is making the default value of `--use-new-uploader` option in the `do-upload` command `True` (so we use the new uploader by default).

But it bothered me taht the upload logic - which is the same between new and old upload endpoints, differing only in the upload sender - was in a module  `services.legacy_upload`. So I moved all files to `services.upload` and put the legacy upload sender in `services.upload.legacy_upload_sender`, and that is as of now the only truly legacy thing in the odecov-cli upload service process.