Update CI workflows.
We have setup a workflow for releasing codecov-cli. Currently it runs on every push to master. If you look at the releases page there are MANY releases being created as drafts. That polutes the visibility of true releases. Furthermore, the only asset we have being built works only on Linux distros, and we'd like to provide something for macos and windows users as well.

We also have redundant CI jobs for PRs and master on push.

To address all those concerns, first we are updating the CI workflows. `push_flow.yml` will run the lint-build-test steps for every push we have. This also uploades test coverage to codecov. `release_flow.yml` takes care of building new releases. It only runs when a new TAG is pushed, starting with `v` (for versions).

To help us use the new release flow a new command is begin added to Makefile, the `tag.release` command. It should be used as `make tag.release version=v0.1.0`, for example. It takes care of tagging a new release and uploading it to GitHub, hopefully triggering the release_flow as well.

The changes for `release_flow.yml` were inspired by [this tutorial](https://data-dive.com/multi-os-deployment-in-cloud-using-pyinstaller-and-github-actions/)