Scaffold runner system for label analysis
Label analysis will eventually support multiple languages, testing frameworks, etc. For that we will need different runners.
Currently we only have `PythonStandardRunner`, and that was hard-coded.

The changes here scaffold a runner system that is similar to the plugin system for upload.
We can extend it and eventually let users have their own runners.
Letting users have their own runners feels particularly important from a safety pov,
users can be certain that the runner is safe to use because they created it.
But of course we also need to provide useful runners for the broad cases. It will be easier for uses to trust our runners once the CLI goes open source.