Add options to PythonStandardRunner and make it safer
* Adds config options to PythonStandardRunner.
They can be configured via the yaml that will be used with the CLI.
Currently options include only collect options.

* Make the python runner execute pytest directly from script, not from subcommand.
To try and make the python runner as safe as possible we are calling pytest directly from the script.
Users can still pass in options, but it's less obvious how to get an attack from that.
At least you can't specify any command to run.