Introduce DAN Runner
DAN (Do Anything Now), inspired from the chatGPT jailbreakers, is a runner that simply does what the user instructs.

We are adding this because we anticipate having hundreds of different scenarios to deal with,
and many different commands. You can think of this as having a custom runner without having to create one.
It is a very power "override all commands" options, with no safety checks.

We will depend heavily on documentation and user-instructing for using this.
The idea is to give the user the power of what will be run when collecting tests and doing something with them.