Don't send unecessary files in static analysis (#130)
You can see that we were already calculating `files_that_need_upload` in static analysis but still iterating over all the files in the response from codecov. This means we were effectively uploading all files everytime. This is unecessary and takes a long time and consumes more data than we need. Plus it was causing issues with the TTL of the pre-signed PUT as well.

The fix is very obvious, just iterate over the files that actually need to be updated.

I also took this opportunity to add some tests to static analysis (coverage is terrible in this command)
THe new requirement is to test the async function in the service.
Maybe it should be in a dedicated requirements.txt inside `tests`?...