Let users pass config to CLI
We have an option in the cli to allow users to pass config to it.
Config controls things like plugins and runners.

Until this point we didn't have to use it, and it was actually broken.
The issue was that `yaml.safe_load` expects a strem (e.g. file contents), not the file name.
So this fixes that issue, and also adds test to make sure the logging is correct.

I decided to move the loading of the config to a helper function to make it easier to test.
Turns out you can't just explicitly call `cli` function because it's wrapped in so much click stuff.
Having it in a separate place makes it easier to test the actual config too.