Make PythonStandardRunner include curr dir by default
Turns out that when you package the CLI in PyInstaller, if you don't have the curr dir in PATH
there's a high probability that pytest will fail because of the pytest plugins (it won't find them)

So it's better to add the current dir by default to avoid issues.

On top of that also adding `params` in the runner interface so we can log the params of the selected runner.