Run pytest in a subprocess using multiprocessing (#139)
This change is because when running label-analysis in the sentry test suite it failed,
with an error saying that a module already imported cannot be rewritten.

Pytest in its documentation about calling it from python has a note saying:

> Calling pytest.main() will result in importing your tests and any modules that they import.
> Due to the caching mechanism of python’s import system, making subsequent calls to pytest.main() from the same process will not reflect changes to those files between the calls.
> For this reason, making multiple calls to pytest.main() from the same process (in order to re-run tests, for example) is not recommended.

And that is exactly what we were doing, calling `pytest.main` twice: collecting tests, and then executing them.

I thought we'd have to go back to `subprocess.run`, which I really didn't want to do because the move to calling pytest from Python directly was *exactly* so we don't have to use `subprocess.run` (a more dangerous alternative, especially with options from the user).

Luckly the multiprocessing lib in its introduction states:
> The multiprocessing package offers both local and remote concurrency, effectively side-stepping the Global Interpreter Lock by __using subprocesses__ instead of threads.

Et voilá, a viable alternative was born.
These changes move the actual execution of `pytest` to `_execute_pytest_subprocess`, and get the response back using a `multiprocessing.Queue`. The interfaces for `collect_tests` and `process_labelanalysis_result` are left unchanged.

Exciting long term possibilities for this change are paralelization of test execution (`process_labelanalysis_result`) if we simply spaw more processes and divide the list of labels among the different processes. That is left for a future change, though.

The question everyone is asking:
Isn't it easier to just use `subprocess.run` though? Isn't this the same thing as doing that?

Yes, that's easier. But we decided to not use it for a reason - it's potentially dangerous from the customer perspective.

No, this is not the same, because only the `_execute_pytest_subprocess` function is executed in a subprocess, which is very explicitly and DOES NOT accept user options that can easily change what program is running.