Improve logging
Trying to strike the right balance between verbose and non-verbose mode in the CLI.

Particularly for the upload command it's been not satisfactory.
The list of files being uploaded as coverage is now listed in non-verbose mode.

Plus there's a positive feedback now that the process (command) finished - for ALL commands.
So you know that at least the CLI did its thing, even though there might still be errors.

And again for the upload command, separate better the response from codecov-api
and the response from the PUT request afterwards.
To help debugging and all that.