Encode PUT data before sending POST request to Codecov
When testing ATS in the sentry fork I have 3 out of 5 uploads timing out when pushing to storage.

I noticed that we are generating the PUT payload *after* receiving the response from Codecov.
Generating PUT payload includes compressing the report file, that might be pretty big,
and considering that the pre-signed put URL has a TTL of 10s this time might be influential.

So now we are just prepping the data completely before starting the upload process,
in the hopes that this will save us enough time to get all uploads into storage in time.