Always feed pytest at least 1 label
We currently have this bug when all tests can be skipped,
then `all_labels` is empty, and we give pytest no labels to execute.
This means it will then collect and execute ALL tests in the suite.

The ideal solution is to bail out and do an empty upload.
However, while empty upload is not available, we are going to select 1 test as random,
and run that single random test label.