Removed TypedDict
TypedDict became part of typing in Python 3.8
We are interested in supporting the CLI in earlier versions of Python 3, so removing it.
It's a bit hacky, but works.
I was careful to make sure we are casting to the new types so we can use the properties if we choose to do so.