Make sure CLI supports multiple Python versions
As the CLI is used by more people we want to make sure it supports some Python versions.
In this case the same ones supported by `codecov/shared`: 3.7 - 3.10

These changes run the build-test-upload steps for those Python versions.
Fix some tests failing for python 3.7 and add status badge for the build steps.

One interesting thing to notice is that the unittest library doesn't seem to work so well with
asyncio in Python 3.7