Enable other runners (DAN runner)
We had created a DAN runner previously, but there was no option in the label analysis
command to actually select different runners.

Another issue is that thwe return of `collect_tests` in a runner needs to be a list,
and the DAN runner was not returning that.

The changes here should make the DAN runner a viable alternative to the
PythonStandardRunner, in case we need that.