Enable custom runners
The trickiest part of ATS is the runner system.
Every customer will have their own test process, and we can't hope to
accomodate all of them. While we will try to provider the most used
runners for them, giving them the option to decide what runner they'll use
is a good idea.

These changes enable the dynamic import of runners.
It works by defining the runner in the yaml config.