Create plugin to compress ATS reports (#154)
Recently we've noticed that JSON reports with labels - the one format accepted by ATS -
are simply too big. It is likely that we are experiencing Out Of Memory errors when
trying to process them.

To put it in perspective, below is a comparison between reports with no labels and
with labels for the worker repo full test suite.
```
  918951 bytes (897Kb) May 15 19:45 coverage.xml # this is the regular coverage with no labels
26607438 bytes (25Mb)  May 15 19:48 label.coverage.json # this is all labels, stock
```

To avoid processing issues and save bandwidth and all that we will be compressing these reports.
The compression idea is simple: we build a mapping label --> int that will substitute the label itself
in the report (usually large-ish strings) for a number in the "contexts" of each file.
We then include the reverse mapping in the report (int --> label) so that we can easily parse
the compressed report when processing it.

This is done via the new `compress_pycoverage_contexts` plugin that can be used when uploading
with the CLI. Notice that the process is transparent for the user.
The results are promissing. Compressing the worker repo (25Mb) yields a new report with 801Kb.

About the new requirements.
These libs allow streaming large files (smart_open) and stream-parsing the JSON (ijson)
This means we don't ever load the full big report in memory when compressing it,
avoiding the Out Of Memory error in the plugin as well.