Patch import errors (the return of subprocess.run) (#172)
Previously (around commit 8c5b289c90a6be8f53eb13babb2a19ef24e1b6ab) we had removed `subprocess.run` from the
`PythonStandardRunner` to make it extra safe. Instead the default way of using it was calling `pytest` directly
from Python.

The thing is that Python import process when running pytest from Python can get messy. Because imported modules
are cached. Out of the blue we started seeing "pytest.PytestAssertRewriteWarning" errors.
The most likely cause is that some modules are imported by the CLI, and when the process that runs Pytest is created
it inherits some of the parent's process resources ([because it's forked](https://docs.python.org/3/library/multiprocessing.html#contexts-and-start-methods)). Yes I tried with "spawn" instead of "fork" but that didn't work either /:

However `subprocess.run` (which probably works differently under the hood?) does work.
It's better to have something slightly more dangerous but that is more reliable in working, than continue to try and make
the safer option work (which (1) is starting to feel hacky and (2) I am out of ideas and we want to ship this feature rather
soon). So the logical choice is to make `subprocess.run` the default way of running pytest. It's also way simpler.

Nevertheless, in the process of trying to make the (now called) strict option work I made some improvement to it in terms
of failing when the subprocess fails. Before it was gettign stuck trying to read from the queue. Now it fails right away.

Ans yes, instead of just throwing away the previous approach we're now calling it "strict mode".