Add dry-run to labelanalysis
This change comes as a request of sentry infra team to increaase our chances of merging ATS into getsentry initially.
The `--dry-run` option simply stops label analysis after collecting labels and prints them in the terminal. Yes, there are
other ways that you can achieve the same result (e.g. custom runner), but this is more straight forward.
It might be marginally useful when setting up or building confidence of label analysis for customers interested in the feature,
so why not?

I decided to add a brief explanation to what each label group is, since we have that opportunity here.