Fix label recalculation
When Codecov doesn't return the absent labels we were re-calculating only the absent labels.
The issue is that, becaue Codecov didn't have them in the first place they couldn't filter out labels that might have been changed in the current commit. And we (CLI) weren't doing that either.

That's an issue because pytest fails if you tell it to run a label that doesn't exist.
So if we get the payload back with no absent labels we have to recalculate everything.