Enable ATS in the CLI
Changing tests that used the `-v` option (invoking the CLI) to instead use the fixture.
This prevents tests from messing up the global env across tests when running pytest using ATS.
As we've been discovering recently, it's a big mess sometimes.

The removal of "MASKED" from some git tests is because it was throwing errors in CI. Maybe it's a python 3.11.x thing...
But I suppose they changed when the removal of secrets run. I changed for bogus secrets.

THe git services was also an error, so I had to put them in a list.