Retry requests on `ConnectionError` and `Timeout`
Adds a retry logic to `send_post_request` and `send_put_request` to retry on ConnectionError and Timeout.
Also makes the legacy upload sender use those functions.

There might be a better way to do that without the custom logic, but the `Retry` option I saw you specify retry error codes,
and I'm not sure ECONNRESET gives an error code. But apparently it is a `requests.exceptions.COnnectionError`, so...

Fixes uploader #537