feat: add support for options when running tests (#230)
There are options that only make sense when running tests (i.e. reporting),
and people use them. So we need to support those as well.

Still no checks there, so the user uses them at their own risk.
Except for the `--cov` option that we are checking to avoid issues, because we do set that one.
So I added a check for it, and we emit a warning message if we find it, but
make no changes.

closes codecov/engineering-team#191