feat: validate commit SHAs
This was motivated by the fact that api returns 404 if your label-analysis request
doesn't include a `base_commit` or a `head_commit`. I don't mean the key must be there,
the value needs to be an accurate value, or it will return a puzzling 404 "Not Found".

That is terrible user experience, and even I needed quite some time to figure it out.
So the best thing (I think) is cutting the evil in its root and making sure the user
gives us a valid SHA.

These changes introduce `validate_commit_sha` to `static-analysis` and `label-analysis` commands.
They make sure the user gives us a string that is 40 chars long and consists of only letters and
numbers. So should at least look like a SHA1 string.