chore: Clean pytest output on error
THe error message if pytest fails to run tests (because of ERROR or FAILURE in tests) was terrible.
There was a huge wall of text with the pytest command and the user had to scroll way up to see anything.
Most of the time they'd read "Pytest did not run correctly" and assume something was wrong with label-analysis

THe present changes address this problem by making the CLI output minimal in this scenario.
I also added hints for the user that label-analysis is probably fine and to check pytest output and
the options used