feat: warn users if coverage report is older than changes
Codecov is only as good as the data you give it.
THis is also true for patch coverage: if the report we are comparing against is old we can't give accurate data.
So mitigate a little this problem patch-coverage now emits a warning if the local report files are older than the
code changes