fix: JS analyser do-statement
We had a call with a customer trying ATS today and they had an error in the JS analyser.
The error was "first_statement referenced before assignment".
This would happen if the first statement in a `do-while` block was not a `statement_block` or `expression_statement`
(I added this case in `sample_004.js` to test it)