chore: add default config
Adding `codecov.yml` as the default config file the CLI will look for.
It's easy for users to forget to add that in the command, and have it misconfigured and then be like
"why is this not working?".

It's also likely that they'll put the config in the codecov.yml where the rest of they codecov config already lives.

context codecov/engineering-team#397