fix: make --dry-run useful
This makes the --dry-run in label-analysis actually do something useful.
It prints the tests you need to run _with_ the tool options to make ats run correctly.

In the case of pytest this means that it prints `--cov-context=test` (we need this to get labels)
and then the list of labels to run, spearated by spaces.

I couldn't find a way to actually export an ENV variable safely so I'm just printing to stdout
The small problem is that there's other output with it, so I decided to add "ATS_TESTS_TO_RUN="
to make grep-ing easier. Just for good measure throw it in a file as well.

So in thory if you get ATS_TESTS_TO_RUN from your CI you can do
`pytest <YOUR NORMAL OPTIONS> <ATS_TESTS_TO_RUN>` and that should work to run all the ATS tests.

closes codecov/engineering-team#470