fix: make max_conditional iterative
context: codecov/codecov-cli#243

User ran into `RecursionError: maximum recursion depth exceeded` running
`_get_max_nested_conditional`.

By removing the recursion from the function we hopefully fix this issue.
So now it's using text-book definition of BFS to go about finding the max depth.

Notice we do have to go through all nodes of the function because we don't know if a node that is not
of type conditional doesn't have another conditional inside of it (think block nodes, for example).
And we only increase the level when we pass through one conditional, not when we go to the next child.

close codecov/codecov-cli#243