fix: Fix python analysis for function docstrings
context: https://l.codecov.dev/hfGkNI

This is a fix for a bug that I stumbled upon today.
After much investigation it boils down to this:
    - treesitter count's docstrings as strings, not comments (which is to be expected I suppose)
    - pytest counts them as comments. They won't appear in the coverage report (if function docstring)
    - if you select a docstring as the line of surety ancestorship, because it's not in the report, it has
    no labels associated with it, so ATS selects no tests for that line

We were doing that (I believe) and selected a function docstring as the line in base to look for tests.
No tests were found, and so ATS passed, but all tests failed.

These changes fix this behavior by doing 2 things:
1. Making sure function docstrings are not in `statements` on the static analysis
2. Making sure we don't select function docstrings as the line of surety ancestorship for the actual
fist line of code in a funciton. It should be `null` instead.