Update --dry-run output
related to codecov/engineering-team#555

These changes update the output of `--dry-run` option so that:
1. Users can get a list of tests that are being skipped
2. Users have the option to parse the output from JSON

> ⚠️
>
>  The JSON output doesn't include the runner options in the tests to run list.
>  Instead there's the 'runner_options' key with that info.
>  This is to make sharding the list of tests easier

It's considerably more output, but given that we don't have clear guidelines on the
best way to do things I think it's better to provide all options.
Eventually adoption will converge to some way or another and we prioritize that.