chore: New release process
This proposed release process is heavily inspired in the worker self-hosted release.
My beef with the current release process is having to create a PR that updates the version number,
AND then manually creating a release.

This proposed workflow will have you manually trigger the `create_release_pr` action with the
version number you want. That will create a PR for you that you can merge yourself.
This PR will then trigger a new release to be created, and the release assets will be generated from that.

THe VERSION file is so we can re-use the existing "create PR" workflow as is.