chore: handle internal `send_single_upload_put` exceptions
Instead of failing (and raising) in the case of a request exception (in `send_single_upload_put`)
we will now catch the exception, emit warning, and return safely.

This is important because the aggregate of all tasks WILL fail if one of the requests fail.
Also notice that `httpx.HTTPError` is the base error class for a number of request-related errors,
so this covers many of the failure possibilities here