chore: rename PythonStandardRunner
This is lower priority, but it has been pointed out to me that `PythonStandardRunner`
is not accurate. There are other test runners available in python after all.

So these changes simple rename it to be `PytestStandardRunner` instead.
Include changes to default options to the `--runner` option and the config key
for the runner. I made that to be backwards compatible tho.