chore: Fix the dry-run output (again)
This is an evolution on the update of dry-run output that was recently
done. That was done in a rush and there have been new learnings since.

The biggest learning is that all log messages go to stderr,
but the dry-run output goes to stdout.
(the reason I didn't know this before is that the test runner
mixes both by default, and I just assumed that everything was going
to stdout. Luckly it's not)

Also I've learned since that the correct way of escaping strings for
shell is actually singlequotes. AND that json is way easier to parse
in the shell, and having everything is terrible

I also learned about tee, that allows you to route stdout from command
to multiple places (including other files).

So now there are explicit format options to the output, and you always
get that in stdout. And only that (in the format you specify).
Default is JSON.

If you want to route it to other files as well as stdout that's fine too.
Up to you, be happy.