fix: Make static analysis handle a large number of uploads
This is more a fix around error handling and updating information to users than actual performance.

**Error Handling**

In terms of error handling, consider that the number of uploads is too big.
We "start" all of them at the same time, but the number of connections is limited.

For uploads you stringify the data and try to grab a connection from the pool to upload.
Then you need to _actually_ upload (e.g. write data to the pipe)

Some timeouts can exist in this process:
* `PoolTimeout` - waiting too long for a connection to become available
* `ReadTimeout` - waiting too long for a response back
(these are the most important ones)

Given that we _will_ have thousands of files to upload it _will_ take a long time,
but that's alright. So we can simply ignore those timeouts.

Then we need to balance the number of connections available.
Through some empirical testing 100 is too much. If you have many big files
trying to upload at the same time the COU time of each is too small, so it takes
too long for the file to finish upload. This throws a 500 error from GCS.

So it's important that we don't try to upload too many files at the same time.
With 20 connections I was able to upload ~6600 files with no errors.

**Update user info**

The progress bar was broken.
It was updating when the single upload was started, not when it finished.
To fix that I just made it update when it finishes each upload.

Now we can properly see that it takes forever :D