chore: improve label analysis log if failed to calculate
If Codecov fails to calculate label analysis (with an actual error),
the log line is borderline unusable, because it cointains too much text.

So these changes restrict the amount of info shown to include:
`head_commit`, `base_commit` and `external_id`. This info should allow
users to quickly check if the comparison is between the commits they expect,
and give them useful info to pass to the support team, so we can learn what happened.

Unfortunatelly at this points the errors are also useless. THey only state
"failed to calculate".