Add test-file and disable-test-search options to CLI
The test-file and disable-test-search options are added to
the global upload options that are used in the do-upload and
upload-process commands.

The test-file option lets users specify specific testing result
files to upload using the same format as the --file option.

The disable-test-search option does the same thing as the --disable-search
option but for testing result files

Signed-off-by: joseph-sentry <joseph.sawaya@sentry.io>