feat: surface label-analysis fallbacks
label-analysis was build to be dependable. Even if the calculation fails (which is rare)
we have many fallback systems in place, so that you can add the label analysis step before
tests and make sure the tests will run.
The fallbacks come with a price, though: running all tests.

One thing we don't do so well is surfacing when label-analysis fallback in a way customers can
use to help build their confidence in label-analysis.

We do have logs that point out that something was not successful and that we tried to fallback.
Sentry specifically requested that we also surface this information in the dry-run output.

These changes to that, letting users check the dry-run output if label-analysis had to
fallback, and if so, what was the reason for it.

closes codecov/engineering-team#711