feat: let users specify pytest command
context: codecov/engineering-team#715

Apparently some users don't have a virtualenv or don't install packages globaly.
In this case running `python -m pytest` doesn't work because pytest is not accessible to the
global python.

These changes let users specify a different command to run, so that they can tell
the PytestStandardRunner where to find `pytest`.

In terms of security it's not more dangerous than having a DAN runner, so should be ok.