chore: remove smart-open
We want to add codecov-cli to getsentry/pypi.
It was asked that we drop the smart-open dependency.

I went back and checked that smart-open is really used for dealing with
files in remote storage and decompressing things under-the-hood.
We don't need any of that, and are well served by the default open function.

So it's fine to drop it.