chore: rename 'pytest_command' to 'python_path' (#344)
We add 'pytest_command' so that users can specify the pytest command to use.
That is still confusing for users to setup the config. Plus people are more used
to give a python path to execute.

So I think it's good idea to replace 'pytest_command' to the simpler 'python_path'.