feat: dynamic runner parameters (#345)
context: codecov/engineering-team#407

Consider this a 1st iteration of this feature.
A compromise between "this is easy to implement and helpful to users"
and "users don't have to think to use this".

The benefit is that now you can add dynamic params to your runners,
the downside is that you need to know the runner configuration AND
format these params accordingly.

Otherwise it would be very hard to figure out exactly what parameter
group a given parameter is. So, compromise.

More details on whoe to actually use the feature are documented in
the code.