feat: support new tokenless protocol (#447)
* feat: add new tokenless support

We no longer have to send the X-Tokenless header
anymore.

Signed-off-by: joseph-sentry <joseph.sawaya@sentry.io>

* fix: remove github API call from tokenless

we no longer want to make the github API call to check if the current
ref we are uploading from is a PR to determine what to set the tokenless
header to. It's no longer necessary and all that is needed for tokenless
is to pass the correct  branch name (containing a ':') to the create
commit step so that tokenless works. The CLI will automatically check if
the TOKENLESS env var has been set by the action. The action will set
this env var when it detects it's running on a fork.

Signed-off-by: joseph-sentry <joseph.sawaya@sentry.io>

* fix: address feedback

* fix: add clarifying comment

Signed-off-by: joseph-sentry <joseph.sawaya@sentry.io>

* test: fix tests

* chore: make lint

Signed-off-by: joseph-sentry <joseph.sawaya@sentry.io>

* chore: make lint

---------

Signed-off-by: joseph-sentry <joseph.sawaya@sentry.io>