🐛 Use config-set token as CLI default fallback
Just like in the old days... Before this patch, the v4-related rewrite
of the Codecov uploader CLI lost the ability to use the token set in
the config file [[1]]. This change fixes the regression recovering the
original behavior.

The bug has been discovered during the re-configuration of
``codecov/codecov-action`` in ``pytest-dev/pytest``. We use the
clear-text token to improve stability of uploads in PRs from forks so
it makes sense to keep it in the config rather than in the GitHub
Actions CI/CD definition file.

[1]: https://docs.codecov.com/docs/codecovyml-reference#codecovtoken