fix: make default search case insensitive
when looking for files in the default patterns search, it's better to
be flexible and not care about case sensitive matching. For example,
certain frameworks output their test results with the word "test" in
all upper case and we shouldn't need to specify both "test" and "TEST"
to match those files.