fix: len(None) bug
the bug was the the NoVersioningSystem class was not defining the
list_relevant_files method, which meant it as falling back to the
VersioningSystemInterface definition which returns None.

This commit make the VersioningSystemInterface an ABC and makes all
the relevant methods abstract, then implements the methods on the
VersioningSystem implementations.