Test on all OSs (#667)
This adds Windows and MacOS to our CI test matrix.

It also changes quite a lot of internal functionality and tests to use `Path.as_posix()` which always prints paths with forward slashes.
This was primarily done to make a ton of tests work on Windows, but IMO it is also a good idea to normalize paths internally to use forward slashes.